+++ lib/components/decorated_navbar.dart
@@ -6,7 +6,7 @@ class DecoratedNavBar extends StatelessWidget {
     super.key,
     this.decoration = const NavBarDecoration(),
     ImageFilter? filter,
-    this.opacity = 1.0,
+    this.opacity = 1,
     this.height = kBottomNavigationBarHeight,
   }) : filter = filter ?? ImageFilter.blur(sigmaX: 3, sigmaY: 3);
 
@@ -17,14 +17,24 @@ class DecoratedNavBar extends StatelessWidget {
   final double height;
 
   @override
-  Widget build(BuildContext context) => ClipRRect(
-        borderRadius:
-            decoration.borderRadius as BorderRadius? ?? BorderRadius.zero,
-        child: BackdropFilter(
-          filter: filter,
-          child: DecoratedBox(
+  Widget build(BuildContext context) => Stack(
+        children: [
+          if (opacity < 1)
+            Positioned.fill(
+              child: ClipRect(
+                child: BackdropFilter(
+                  filter: decoration.filter ?? filter,
+                  child: Container(
+                    color: Colors.transparent,
+                  ),
+                ),
+              ),
+            ),
+          DecoratedBox(
             decoration: decoration.copyWith(
-              color: decoration.color?.withOpacity(opacity),
+              color: opacity != 1
+                  ? decoration.color?.withOpacity(opacity)
+                  : decoration.color,
             ),
             child: SafeArea(
               top: false,
@@ -37,6 +47,6 @@ class DecoratedNavBar extends StatelessWidget {
               ),
             ),
           ),
-        ),
+        ],
       );
 }